fastapi/requirements.txt
fastapi~=0.115.5
python-multipart
uvicorn~=0.32.0
jinja2==3.1.4

starlette~=0.42.0
dotenv~=0.9.9
python-dotenv~=1.0.1
gspread~=6.2.0
oauth2client~=4.1.3
pydantic~=2.9.2

premailer
reportlab
pandas

passlib[bcrypt]
python-jose